[CHG] replace arrayview with my own view

+++ common_audio/signal_processing/dot_product_with_scale.c
@@ -31,5 +31,5 @@ int32_t WebRtcSpl_DotProductWithScale(const int16_t* vector1,
   }
 
   // rtc::saturated_cast<int32_t>(sum);
-  return (int32_t)WEBRTC_SPL_SAT(WEBRTC_SPL_WORD32_MAX, sum, WEBRTC_SPL_WORD32_MIN);
+  return WEBRTC_SPL_SAT(WEBRTC_SPL_WORD32_MAX, sum, WEBRTC_SPL_WORD32_MIN);
 }
